fix: Add type casts to string functions and fix middleware issues
Cast char pointers to unsigned types in isspace/isBlank and isdigit calls to prevent undefined behavior with signed chars; fix typo in doxygen comment; update base64_decode signature to use const char*; remove unused struct definition; streamline request validation guard rail. These changes improve code portability, safety, and compliance with C standards.

+++ lib/http-server-middleware/src/middleware_conditional.c
@@ -207,7 +207,7 @@ static esp_err_t parse_etag_list(const char *header_value, char etags[][HTTPD_MA
     // Parse comma-separated ETags
     while (*ptr && *etag_count < max_etags) {
         // Skip whitespace
-        while (*ptr && isspace(*ptr)) ptr++;
+        while (*ptr && isspace((unsigned int)*ptr)) ptr++;
         
         if (!*ptr) break;
         
@@ -222,10 +222,10 @@ static esp_err_t parse_etag_list(const char *header_value, char etags[][HTTPD_MA
             const char *etag_end = start + len - 1;
 
             // Trim leading whitespace
-            while (etag_start <= etag_end && isspace(*etag_start)) etag_start++;
+            while (etag_start <= etag_end && isspace((unsigned int)*etag_start)) etag_start++;
 
             // Trim trailing whitespace
-            while (etag_end > etag_start && isspace(*etag_end)) etag_end--;
+            while (etag_end > etag_start && isspace((unsigned int)*etag_end)) etag_end--;
 
             if (etag_start > etag_end) {
                 // Skip empty ETags after trimming
